Hotfix

+++ CoinModel.cpp
@@ -20,7 +20,7 @@ QVariant CoinModel::data(const QModelIndex& index, int role) const
     if      (!coinInfosMap)                                                                      return QVariant();
     else if (index.row() < 0 || index.row() >= coinInfosMap->count() || role != Qt::DisplayRole) return QVariant();
 
-    else if (index.column() == 0) return coinInfosMap->keys().at(index.row());
+    else if (index.column() == 0) return (qlonglong)coinInfosMap->keys().at(index.row());
     else if (index.column() == 1) return &(coinInfosMap->values().at(index.row()));
 
     else return QVariant();
